requirements.txt
Django==5.2.4
django-ninja==1.4.3
django-environ==0.12.0
psycopg2-binary==2.9.10
django-ninja==1.4.3
django-ninja-extra==0.30.1
django-ninja-jwt==5.3.7
